+++ src/main/io/transponder_ir.h
@@ -43,7 +43,7 @@ extern const transponderRequirement_t transponderRequirements[TRANSPONDER_PROVID
 
 void transponderInit(void);
 
-uint16_t transponderUpdate(PifTask *p_task);
+uint32_t transponderUpdate(PifTask *p_task);
 void transponderUpdateData(void);
 void transponderTransmitOnce(void);
 void transponderStartRepeating(void);
