+++ src/main/osd/osd.c
@@ -1176,7 +1176,7 @@ bool osdUpdateCheck(timeUs_t currentTimeUs, timeDelta_t currentDeltaTimeUs)
 }
 
 // Called when there is OSD update work to be done
-uint16_t osdUpdate(PifTask *p_task)
+uint32_t osdUpdate(PifTask *p_task)
 {
     static uint16_t osdStateDurationFractionUs[OSD_STATE_COUNT] = { 0 };
     static uint32_t osdElementDurationUs[OSD_ITEM_COUNT] = { 0 };
